Remove CDA file handling and enhance statistics features
This commit eliminates the handling of CDA files in the `UploadSpeaker` method and the associated `CdaConversionService`. The `SpeakerModel` has been updated to remove references to CDA files, and the maximum file size for MP3 uploads has been increased to 700MB.

Additionally, the `StatsController` and `VisitorRepository` have been enhanced to support new methods for retrieving monthly unique visitor data and available years. The UI in `AddSpeaker.cshtml` and `Index.cshtml` has been updated to reflect these changes, including the removal of the conversion modal and the addition of a chart for displaying unique connected devices. CSS styles have also been improved for better layout and appearance.